renamed TradingController to TraderController, moved trader function from GameController to TraderController
