+++ lib/controllers/index.js
@@ -2,7 +2,7 @@ const ClientController = require("./ClientController");
 const FriendController = require("./FriendController");
 const GameController = require("./GameController");
 const MenuController = require("./MenuController");
-const TradingController = require("./TradingController");
+const TraderController = require("./TraderController");
 const BundlesController = require("./BundlesController");
 const LocationController = require("./LocationController");
 const AccountController = require("./AccountController");
@@ -23,7 +23,7 @@ module.exports = {
     ...GameController,
     ...LocationController,
     ...MenuController,
-    ...TradingController,
+    ...TraderController,
     ...WeblauncherController,
     ...RaidController,
     ...ItemController,
